Aysa_Risk_Front/src/app/app.component.spec.ts
/* eslint-disable no-undef */
import { TestBed } from '@angular/core/testing';
// eslint-disable-next-line sort-imports
import { RouterTestingModule } from '@angular/router/testing';
import { AppComponent } from './app.component';

// eslint-disable-next-line no-undef
describe('AppComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [RouterTestingModule],
      declarations: [AppComponent]
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'ng_zorro_theme'`, () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('ng_zorro_theme');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.content span')?.textContent).toContain('ng_zorro_theme app is running!');
  });
});
